Add a new class Net::OpenID::IndirectMessage which is an abstraction layer around an OpenID indirect message using arguments with the openid. prefix.
The new spec requires the ability to enumerate the argument keys
in order to support extensions, so I figured we might as well
go the whole hog and rework the argument handling to be a bit
nicer while at it. (Won't need to duplicate the openid. prefix
all over the place, for example.)

This isn't used yet. It hasn't even been tested. Really just
checking it in for visibility's sake.